+++ src/common/icons/CrossIcon.tsx
