+++ src/common/icons/ExplosionIcon.tsx
