+++ src/common/icons/PauseIcon.tsx
