+++ src/common/icons/PlayIcon.tsx
