+++ src/common/icons/RemoveFromListIcon.tsx
