+++ src/common/icons/SpinnerIcon.tsx
