+++ src/common/icons/TrashIcon.tsx
