--- src/features/tracks/DownArrowIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const DownArrowIcon = ({ className }: { className: string }) => (
-  <svg className={className} width={20} height={20} viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-    <path d="M5 6h10l-5 9-5-9z" />
-  </svg>
-);
