--- src/features/tracks/UpArrowIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const UpArrowIcon = ({ className }: { className: string }) => (
-  <svg className={className} width={20} height={20} viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-    <path d="M15 14H5l5-9 5 9z" />
-  </svg>
-);
